+++ lib/vindi/railtie.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Vindi
+  class Railtie < ::Rails::Railtie
+    # This class allows Rails to load generator configurations automatically
+  end
+end